Bug 882086 - [MMS] [UX] Message thread. Spacing between the last bubble and input field should be 1.5 rem r=julienw
Previously, the spacing after the final message "bubble" was set to an
arbitrary value of 7rem. This patch sets the spacing to the specified
amount, and ensures that the spacing remains consistent as the Compose
field grows and shrinks (resolving "Bug 875362 - [MMS] Cannot scroll to
bottom of thread list as input field grows")

This patch also addresses an issue where the Compose field could grow to
completely cover the message thread. Now, a "sliver" of space is
reserved to maintain a visual affordance to the underlying thread.